dedupe: move to reflink module + stand-alone option parsing
